+++ packages/client/redux/instanceConnection/service.ts
@@ -17,6 +17,13 @@ export function provisionInstanceServer (locationId: string, instanceId?: string
       dispatch(instanceServerProvisioning());
       const token = getState().get('auth').get('authUser').accessToken;
       console.log(`Provisioning instance server for location ${locationId} and instance ${instanceId}`);
+      console.log("Query is")
+      console.log({
+        locationId: locationId,
+        instanceId: instanceId,
+        sceneId: sceneId,
+        token: token
+      })
       const provisionResult = await client.service('instance-provision').find({
         query: {
           locationId: locationId,
@@ -25,6 +32,7 @@ export function provisionInstanceServer (locationId: string, instanceId?: string
           token: token
         }
       });
+      console.log("Provision result is: ")
       console.log(provisionResult);
       if (provisionResult.ipAddress != null && provisionResult.port != null) {
         dispatch(instanceServerProvisioned(provisionResult, locationId, sceneId));
